Use correct length for unicode/emoji strings
Strings with unicode or emoji are not handled properly by default
with python. Glyphs that are 2 or 4 characters wide will have a
`len()` of 1. This causes asterisks at the end of task names to wrap
to the next line and it looks awkward.

If the `wcwidth` module is present on the system, use it to determine
the correct length of the string. This allows the asterisk padding after
task names to be the correct length without wrapping.

Fixes: ansible/ansible#63105

Signed-off-by: Major Hayden <major@redhat.com>